requirements.txt
yfinance==0.2.54
pandas==2.2.3
numpy==2.2.1
python-telegram-bot==22.6
schedule==1.2.2
python-dotenv==1.2.1
